dom script isolated
